backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
sendgrid
python-dotenv
bcrypt
celery
google-api-python-client
google-auth
requests==2.31.0
pandas
pytz
